Extract Request data to separate files and namespace
