+++ src-d/diva-dns/Requests/PostRequest.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using diva_dns.Data;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http.Json;
