+++ src-d/diva-dns/Requests/PutRequest.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Net.Http.Json;
 using System.Text;
 using System.Threading.Tasks;
+using diva_dns.Data;
 
 namespace diva_dns.Requests
 {
@@ -11,11 +12,11 @@ namespace diva_dns.Requests
     {
         private readonly string _url;
         private readonly HttpClient _client;
-        private readonly Data _data;
+        private readonly Data.Data _data;
 
         public HttpResponseMessage? ResponseMessage { get; private set; }
 
-        public PutRequest(HttpClient client, string url, Data data)
+        public PutRequest(HttpClient client, string url, Data.Data data)
         {
             _client = client;
             _url = url;
